Move link command logic out of index.ts into link-project.ts
index.ts had grown into a mix of validation, git/path resolution, and
output formatting for a single command. Consolidate it all into one
linkProject() function (the right seams turned out to be one, not three
— entry derivation and target resolution are the same concern, and
summary formatting was too trivial to isolate), and make index.ts a
pure re-export so there's no proxy function to keep in sync with it.
